+++ src/main/kotlin/dsm/pick2024/domain/timetable/port/in/QueryDayTimetableUseCase.kt
@@ -1,6 +1,6 @@
 package dsm.pick2024.domain.timetable.port.`in`
 
-import dsm.pick2024.domain.timetable.presentation.dto.DayTimetableResponse
+import dsm.pick2024.domain.timetable.presentation.dto.response.DayTimetableResponse
 
 interface QueryDayTimetableUseCase {
     fun queryDayTimetable(): DayTimetableResponse
