+++ src/main/kotlin/dsm/pick2024/domain/timetable/port/in/QueryWeekTimetableUseCase.kt
@@ -1,6 +1,6 @@
 package dsm.pick2024.domain.timetable.port.`in`
 
-import dsm.pick2024.domain.timetable.presentation.dto.DayTimetableResponse
+import dsm.pick2024.domain.timetable.presentation.dto.response.DayTimetableResponse
 
 interface QueryWeekTimetableUseCase {
     fun queryWeekTimetable(): List<DayTimetableResponse>
